refactor(frontend): split 7 large components into smaller modules (H-1)
Split 7 oversized components (400+ LOC) into 21 focused files while
maintaining all existing tests and public exports via re-exports.

Splits performed:
- CvSection (427→262) + cv-state.ts (144) + NullableNumberField (37)
- ConfigForm (586→495) + ModelParamsSection (91)
- TuneTab (445→213) + tune-config-utils.ts (45) + TuneEvaluationSection (229)
- SearchSpaceTable (607→301) + search-space-utils.ts (94) + SearchSpaceRow (271)
- JobDetail (540→452) + ConfigTreeView (79)
- ResultsPanel (763→305) + ResultsRunningView (166) + ResultsCompletedView (288) + LiveTrialChart (51)
- DataPanel (739→206) + DataSourceSection (141) + ColumnSettingsSection (185) + useDataPanel hook (369)

Not split (within limits, high cohesion):
- field-renderers.tsx (578) — recursive structure, circular dep risk
- BlockedGroupKFoldEditor.tsx (474) — shared state, no clear boundary
- ModelPanel.tsx (434) — within limit, high cohesion

All 1157 tests pass, build succeeds, Biome clean.
