+++ frontend/src/components/jobs/JobDetail.tsx
@@ -1,13 +1,5 @@
 import { useQuery, useQueryClient } from "@tanstack/react-query";
-import {
-  ArrowRight,
-  ChevronDown,
-  ChevronRight,
-  Download,
-  RotateCcw,
-  Trash2,
-  X,
-} from "lucide-react";
+import { ArrowRight, Download, RotateCcw, Trash2, X } from "lucide-react";
 import { useCallback, useEffect, useRef, useState } from "react";
 import { useNavigate } from "react-router-dom";
 import { toast } from "sonner";
@@ -31,6 +23,7 @@ import {
 import { Progress } from "@/components/ui/progress";
 import { formatElapsed } from "@/lib/utils";
 import { CompletedContent } from "./CompletedContent";
+import { ConfigTreeView } from "./ConfigTreeView";
 import { DeleteDialog } from "./DeleteDialog";
 import { ExportDialog } from "./ExportDialog";
 
@@ -455,86 +448,5 @@ function LogDialog({
         </pre>
       </DialogContent>
     </Dialog>
-  );
-}
-
-/* ------------------------------------------------------------------ */
-/*  Config Tree View                                                   */
-/* ------------------------------------------------------------------ */
-
-function ConfigTreeView({ data }: { data: unknown }) {
-  if (data == null) {
-    return <span className="text-muted-foreground italic">null</span>;
-  }
-
-  if (typeof data !== "object") {
-    return <span className="font-mono">{String(data)}</span>;
-  }
-
-  if (Array.isArray(data)) {
-    if (data.length === 0) {
-      return <span className="font-mono text-muted-foreground">[]</span>;
-    }
-    return (
-      <ul className="space-y-0.5">
-        {data.map((item, idx) => (
-          <li key={`${idx}`} className="flex items-start gap-1">
-            <span className="text-muted-foreground select-none">-</span>
-            <ConfigTreeView data={item} />
-          </li>
-        ))}
-      </ul>
-    );
-  }
-
-  const entries = Object.entries(data as Record<string, unknown>);
-  if (entries.length === 0) {
-    return <span className="font-mono text-muted-foreground">{"{}"}</span>;
-  }
-
-  return (
-    <ul className="space-y-0.5">
-      {entries.map(([key, value]) => (
-        <ConfigTreeNode key={key} label={key} value={value} />
-      ))}
-    </ul>
-  );
-}
-
-function ConfigTreeNode({ label, value }: { label: string; value: unknown }) {
-  const isExpandable =
-    value != null && typeof value === "object" && Object.keys(value).length > 0;
-  const [expanded, setExpanded] = useState(false);
-
-  if (!isExpandable) {
-    return (
-      <li className="flex items-start gap-1">
-        <span className="w-3.5 shrink-0" />
-        <span className="font-semibold">{label}:</span>{" "}
-        <ConfigTreeView data={value} />
-      </li>
-    );
-  }
-
-  return (
-    <li>
-      <button
-        type="button"
-        className="flex items-start gap-0.5 hover:bg-accent/50 rounded px-0.5 -ml-0.5"
-        onClick={() => setExpanded((prev) => !prev)}
-      >
-        {expanded ? (
-          <ChevronDown className="mt-0.5 h-3 w-3 shrink-0" />
-        ) : (
-          <ChevronRight className="mt-0.5 h-3 w-3 shrink-0" />
-        )}
-        <span className="font-semibold">{label}</span>
-      </button>
-      {expanded && (
-        <div className="ml-4 border-l pl-2">
-          <ConfigTreeView data={value} />
-        </div>
-      )}
-    </li>
   );
 }
