+++ frontend/src/components/workspace/ConfigForm.tsx
@@ -1,4 +1,4 @@
-import { useCallback, useEffect, useMemo, useState } from "react";
+import { useCallback, useEffect, useMemo } from "react";
 import {
   Accordion,
   AccordionContent,
@@ -22,12 +22,12 @@ import {
   type SchemaProperty,
   setNestedValue,
 } from "./config-utils";
-import { DynParam } from "./DynParam";
 import { FeatureWeightsEditor } from "./FeatureWeightsEditor";
 import { FormField } from "./FormField";
 import { renderField } from "./field-renderers";
 import { KeyValueEditor } from "./KeyValueEditor";
 import { MetricsChips } from "./MetricsChips";
+import { ModelParamsSection } from "./ModelParamsSection";
 import { NumberInput } from "./NumberInput";
 
 interface ConfigFormProps {
@@ -491,96 +491,5 @@ export function ConfigForm({
         )}
       </Accordion>
     </div>
-  );
-}
-
-// --- Model Params with Essential / Advanced split ---
-
-const ESSENTIAL_PARAM_KEYS = new Set([
-  "objective",
-  "metric",
-  "n_estimators",
-  "learning_rate",
-  "max_depth",
-  "num_leaves",
-]);
-
-function ModelParamsSection({
-  hints,
-  getValueForHint,
-  handleHintChange,
-  getOptionsForHint,
-  shouldShowField,
-  precisionAtKValue,
-  onPrecisionAtKChange,
-}: {
-  hints: import("@/api/types").ParameterHint[];
-  getValueForHint: (hint: import("@/api/types").ParameterHint) => unknown;
-  handleHintChange: (
-    hint: import("@/api/types").ParameterHint,
-    value: unknown,
-  ) => void;
-  getOptionsForHint: (hint: import("@/api/types").ParameterHint) => string[];
-  shouldShowField: (key: string) => boolean;
-  precisionAtKValue?: number;
-  onPrecisionAtKChange?: (k: number) => void;
-}) {
-  const [showAdvanced, setShowAdvanced] = useState(false);
-
-  const essential = hints.filter((h) => ESSENTIAL_PARAM_KEYS.has(h.key));
-  const advanced = hints.filter((h) => !ESSENTIAL_PARAM_KEYS.has(h.key));
-
-  return (
-    <>
-      {essential.map((hint) => (
-        <DynParam
-          key={hint.key}
-          hint={hint}
-          value={getValueForHint(hint)}
-          onChange={(v) => handleHintChange(hint, v)}
-          options={getOptionsForHint(hint)}
-          visible={shouldShowField(hint.key)}
-          precisionAtKValue={
-            hint.kind === "model_metric" ? precisionAtKValue : undefined
-          }
-          onPrecisionAtKChange={
-            hint.kind === "model_metric" ? onPrecisionAtKChange : undefined
-          }
-        />
-      ))}
-      {advanced.length > 0 && (
-        <>
-          <button
-            type="button"
-            className="mt-1 mb-1 text-xs text-muted-foreground hover:text-foreground transition-colors"
-            onClick={() => setShowAdvanced((v) => !v)}
-            data-testid="toggle-advanced-params"
-          >
-            {showAdvanced
-              ? `▾ Hide advanced (${advanced.length})`
-              : `▸ Show advanced (${advanced.length})`}
-          </button>
-          {showAdvanced &&
-            advanced.map((hint) => (
-              <DynParam
-                key={hint.key}
-                hint={hint}
-                value={getValueForHint(hint)}
-                onChange={(v) => handleHintChange(hint, v)}
-                options={getOptionsForHint(hint)}
-                visible={shouldShowField(hint.key)}
-                precisionAtKValue={
-                  hint.kind === "model_metric" ? precisionAtKValue : undefined
-                }
-                onPrecisionAtKChange={
-                  hint.kind === "model_metric"
-                    ? onPrecisionAtKChange
-                    : undefined
-                }
-              />
-            ))}
-        </>
-      )}
-    </>
   );
 }
